+++ mail/proton-mail-common/tests/sidebar_custom_labels.rs
@@ -3,7 +3,7 @@ use crate::common::TestContext;
 use proton_api_core::services::proton::common::RemoteId as ApiRemoteId;
 use proton_api_mail::services::proton::common::{LabelType as ApiLabelType, LabelType};
 use proton_api_mail::services::proton::response_data::Label as ApiLabel;
-use proton_core_common::datatypes::{LabelId, RemoteId};
+use proton_core_common::datatypes::LabelId;
 use proton_mail_common::datatypes::SystemLabelId;
 use proton_mail_common::Sidebar;
 use test_case::test_case;
@@ -12,7 +12,7 @@ use velcro::hash_map;
 mod common;
 
 #[test_case(&[], &[]; "empty")]
-#[test_case(&[(ApiRemoteId::from("foo"), "foo".to_owned(), 42)], &[(RemoteId::from("foo").into(), "foo".to_owned())]; "single")]
+#[test_case(&[(ApiRemoteId::from("foo"), "foo".to_owned(), 42)], &["foo".to_owned()]; "single")]
 #[test_case(&[
     (ApiRemoteId::from("bar"), "bar".to_owned(), 2),
     (ApiRemoteId::from("baz"), "baz".to_owned(), 3),
@@ -20,21 +20,19 @@ mod common;
     (ApiRemoteId::from("titi"), "titi".to_owned(), 5),
     (ApiRemoteId::from("toto"), "toto".to_owned(), 4),
 ], &[
-    (RemoteId::from("foo").into(), "foo".to_owned()),
-    (RemoteId::from("bar").into(), "bar".to_owned()),
-    (RemoteId::from("baz").into(), "baz".to_owned()),
-    (RemoteId::from("toto").into(), "toto".to_owned()),
-    (RemoteId::from("titi").into(), "titi".to_owned())
+    "foo".to_owned(),
+    "bar".to_owned(),
+    "baz".to_owned(),
+    "toto".to_owned(),
+    "titi".to_owned()
 ]; "many")]
 #[tokio::test]
-async fn sidebar_custom_labels(
-    labels: &[(ApiRemoteId, String, u32)],
-    expected: &[(LabelId, String)],
-) {
+async fn sidebar_custom_labels(labels: &[(ApiRemoteId, String, u32)], expected: &[String]) {
     // Setup:
     //   * Setup User:
     //     + Create Custom Folders
     //   * Create Sidebar
+
     let ctx = TestContext::new().await;
     ctx.setup_user(sidebar_test_params(labels)).await;
 
@@ -51,10 +49,7 @@ async fn sidebar_custom_labels(
     let result = sidebar.custom_labels().await.unwrap();
 
     // Tests
-    let result: Vec<_> = result
-        .into_iter()
-        .map(|l| (l.remote_id.unwrap(), l.name))
-        .collect();
+    let result: Vec<_> = result.into_iter().map(|l| l.name).collect();
     assert_eq!(result, expected);
 }